[#767] fix(bff/ml): resolve Schema-000054-Compat step-9 failures (8 tests)
Three test-code bugs (bad UUID literals, int-as-bool pgx encoding,
permanent table FK to temp table) and one real production query bug
(SQLSTATE 42P08 type-deduction conflict on $4/$5 in upsertQ) fixed so
the Schema-000054-Compat step-9 suite exits 0 end-to-end.

Closes #767

+++ services/bff/internal/storage/repository/decks_repo_test.go
@@ -59,7 +59,7 @@ func insertTestDeckCardAsInteger(t *testing.T, db *sql.DB, deckID string, cardID
 	_, err := db.ExecContext(
 		context.Background(),
 		`INSERT INTO deck_cards (deck_id, card_id, quantity, board, from_draft_pick)
-		 VALUES ($1, $2, 1, 'main', $3::boolean)
+		 VALUES ($1, $2, 1, 'main', ($3::int)::boolean)
 		 ON CONFLICT (deck_id, card_id, board) DO NOTHING`,
 		deckID, cardID, fromDraftPickInt,
 	)
